I have made all the internal functions in `SeatingChartViewModel.kt` suspend functions.

+++ app/src/main/java/com/example/myapplication/data/HomeworkLog.kt
@@ -20,7 +20,7 @@ import androidx.room.PrimaryKey
 data class HomeworkLog(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
-    val studentId: Int,
+    val studentId: Long,
     val assignmentName: String,
     val status: String,
     val loggedAt: Long = System.currentTimeMillis(),
